requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
pyyaml>=6.0

